requirements.txt
python-telegram-bot>=20.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
qbittorrent-api>=2023.10.50
pydantic>=2.0.0
pydantic-settings>=2.0.0
lxml>=4.9.0

